New hash and equal comparision functions for double
Previous g_double_hash function simply cast the gdouble value into a
guint. That poses many problems. First, of all values a double can
assume, the cast to guint is well defined only for positive numbers up to
UINT_MAX (see C99 §6.3.1.4). Thus, there isn't even guarantee that the
hash value will be the same for a given input (although such behavior
is highly unlikely). On GCC, casting a negative double value to unsigned
yields 0, what brings us to the second problem: too many values would
have a hash value of zero. With GCC, at least half of the number spaces
of double have the same hash value of zero. But if you consider double has
a floating precision, it is much more than that. It is a very common use
case to have all your double dataset in the range of [0, 1), as is the
output of g_rand_double. So, if I had a thousand numbers generated with
g_rand_double, by the double to unsigned conversion rule, all of them
would also have the same hash value of zero.

This new hash function, inspired by boost::hash_combine (GPL compatible
license), hashes the binary representation of the value (that is unique
for +inf, -inf, and every number except 0), treats the special cases of
+0.0 and -0.0 as the same, and also consider all NaN the same (that have
many different possible binary representations).

The g_double_equal is extended to consider NaN == NaN to be true. Plain
double comparision considers it false, so using it in GHashTable makes
it possible for adding keys that can never be retrieved.