arraylist: add GArrayList
GArrayList is a datastructure that allows you to store data internally as
an array for fast index based access. However, it maintains that array as
an array of GList so that you can also use the data in compatability APIs
as a GList.

This has some additional costs in keeping the array based GList pointers
in sync, but overhead is mitigated in the common cases of small lists.

You can also embed the structure removing malloc overhead for small cases
such as [0..2].